Reorganized the code, splitting the FLask app from the summarizer.
