+++ README.md
@@ -2,17 +2,22 @@
 
 Code to create a summary of an MMIF file, only keeping those annotations that are useful metadata (and reducing the size of the file by one or two orders of magnitude), and making some implicit relations between annotations and anchors in the source explicit.
 
-See `code/run.py` for some of the limitations of the code.
+See `code/summary.py` for some of the limitations of the code.
 
+This code requries Python 3.8 or higher and two modules:
 
+```bash
+$ pip install clams-python==0.5.3
+$ pip install beautifulsoup4
+```
 
 ## Usage
 
-Run the application without using the server by running the summarizer in test mode (using the -t option) on one of the sample input documents: `examples/input-v7.mmif` and `examples/input-v9.mmif`. The first has seven manually created views iews from the following applications: bars-and-tone, slates, audio-segmenter, kaldi, east, tesseract and slate-parser. The second has two views added by the app-spacy-nlp application.
+Run the summary code in stand-alone mode on one of the sample input documents: `examples/input-v7.mmif` and `examples/input-v9.mmif`. The first has seven manually created views iews from the following applications: bars-and-tone, slates, audio-segmenter, kaldi, east, tesseract and slate-parser. The second has two views added by the app-spacy-nlp application.
 
 ```bash
 $ cd code
-$ python run.py -t examples/input-v7.mmif
+$ python summary.py examples/input-v7.mmif
 ```
 
 The output for this should be something like
@@ -43,10 +48,10 @@ The example outputs for both input examples are in the `examples` folder.
 To start a Flask server:
 
 ```bash
-$ python run.py
+$ python app.py
 ```
 
-To use the server (do this from a different terminal than the run.py script, but from the directory with this readme file):
+To use the server from curl (do this from a different terminal than the run.py script, but from the directory with this readme file):
 
 ```bash
 $ curl -X GET http://0.0.0.0:5000/
@@ -54,7 +59,6 @@ $ curl --request POST -d@examples/input-v7.mmif http://0.0.0.0:5000/
 ```
 
 The first gets you the metadata, the second sends off the file `examples/input-v7.mmif` to get summarized.
-
 
 
 ## Docker
@@ -70,10 +74,10 @@ Entering the container and testing the converter:
 
 ```bash
 $ docker run --rm -it mmif-summarizer bash
-$ python run.py -t examples/input-v7.mmif
+$ python summary.py examples/input-v7.mmif
 ```
 
- In real life, the second prompt will be the container prompt, something like `root@f9deb9a6115f:/app#`.
+In real life, the second prompt will be the container prompt, something like `root@f9deb9a6115f:/app#`.
 
 To run the Flask server in a Docker container:
