+++ code/Dockerfile
@@ -1,10 +1,9 @@
-FROM python:3.8-slim-buster
+FROM ghcr.io/clamsproject/clams-python:0.5.3
 
-RUN pip3 install clams-python==0.5.1
 RUN pip3 install beautifulsoup4===4.9.3
 
 WORKDIR ./app
 
 COPY ./ ./
 
-CMD ["python3", "run.py"]
+CMD ["python3", "app.py"]
