+++ code/requirements.txt
@@ -1,2 +1,2 @@
-clams-python==0.5.1
+clams-python==0.5.3
 beautifulsoup4===4.9.3
