+++ code/summary.py
@@ -1,4 +1,4 @@
-"""run.py
+"""MMIF Summarizer
 
 MMIF consumer that creates a summary XML file.
 
@@ -25,117 +25,36 @@ Makes some simplifying assumptions, some of them need to be revisited:
 
 USAGE:
 
-To run just the summarizer on one of the simple sample input documents:
+Run the summarizer on one of the simple example input documents:
 
-$ python run.py -t examples/input-v7.mmif
-$ python run.py -t examples/input-v9.mmif
+$ python summary.py examples/input-v7.mmif
+$ python summary.py examples/input-v9.mmif
 
 The second one is like the first one but also includes spaCy annotations.
 
-To start a Flask server and ping it:
-
-$ python run.py
-$ curl -X GET http://0.0.0.0:5000/
-$ curl -X POST -d@examples/input-v7.mmif http://0.0.0.0:5000/
-$ curl -X POST -d@examples/input-v9.mmif http://0.0.0.0:5000/
-
-See README.md for more details.
-
 TODO:
+
 - entities from spacy are aligned with full document, not tokens therein
 - define summary() to create a summary and then separate the export into XML
   versus JSON
 - run this on a real file with Kaldi and other real annotations:
     Bars-and-tone --> Slate Recognition --> Slate parser mockup -->
     Segmenter --> Kaldi --> EAST --> Tesseract --> SpacyNER
-
 """
 
-import sys
-import json
-import io
+import io, sys
 
-from server import ClamsConsumer, Restifier
-from clams.appmetadata import AppMetadata
 from mmif.serialize import Mmif
-from mmif.vocabulary import DocumentTypes
-from mmif.vocabulary import AnnotationTypes
-from lapps.discriminators import Uri
 
-from utils import type_name, as_xml, write_tag, xml_attribute, xml_data
-from utils import flatten_paths, print_paths
-from utils import get_annotations_from_view, find_matching_tokens
+from utils import as_xml, write_tag, xml_attribute, xml_data
 from graph import Graph
 import names
+from config import GRANULARITY, TRANSCRIPT, KALDI
+from config import TF_PROPS, E_PROPS, VIEW_PROPS, DOC_PROPS
 
 
-VERSION = '0.1.1'
-MMIF_VERSION = '0.4.0'
-MMIF_PYTHON_VERSION = '0.4.6'
-CLAMS_PYTHON_VERSION = '0.5.1'
-
-
-# The name of the Kaldi app, used to select views
-KALDI = 'http://apps.clams.ai/kaldi/'
-
-# Bounding boxes have a time point, but what we are looking for below is to find
-# a start and an end in the video so we manufacture an end point. Set to 1000ms
-# because Tesseract samples every second
-# TODO: this is not used anymore, probably needs to be re-introduced
-MINIMAL_TIMEFRAME_LENGTH = 1000
-
-# When a named entity occurs 20 times we do not want to generate 20 instances of
-# it. If the start of the next entity occurs within the below number of
-# miliseconds after the end of the previous, then it is just added to the
-# previous one. Taking one minute as the default so two mentions in a minute end
-# up being the same instance. This setting can be changed with the 'granularity'
-# parameter.
-# TODO: this seems broken
-GRANULARITY = 1000
-GRANULARITY_HELP = 'maximum interval between two entities in the same cluster'
-
-# The transcript is probably the largest part of the summary, but in some cases
-# it is not needed for the user, in which case we can suppress it from being in
-# the output This setting can be changed with the 'transcript' parameter.
-TRANSCRIPT = True
-TRANSCRIPT_HELP = ''
-
-# Properties used for the summary for various tags
-DOC_PROPS = ('id', 'type', 'location')
-VIEW_PROPS = ('id', 'timestamp', 'app')
-TF_PROPS = ('id', 'start', 'end', 'frameType')
-E_PROPS = ('id', 'group', 'cat', 'tag', 'video-start', 'video-end', 'coordinates')
-
-
-class MmifSummarizer(ClamsConsumer):
-
-    def _consumermetadata(self):
-        self.metadata = \
-            AppMetadata(
-                identifier="https://apps.clams.ai/mmif-summarizer",
-                url='https://github.com/clamsproject/mmif-summarizer',
-                name="MMIF Summarizer",
-                description="Summarize a MMIF file.",
-                mmif_version=MMIF_VERSION,
-                app_version=VERSION,
-                app_license='Apache 2.0',
-                analyzer_version=VERSION,
-                analyzer_license='Apache 2.0')
-        self.metadata.add_input(DocumentTypes.TextDocument, required=False)
-        self.metadata.add_input(AnnotationTypes.TimeFrame, required=False)
-        self.metadata.add_input(AnnotationTypes.BoundingBox, required=False)
-        self.metadata.add_input(AnnotationTypes.Alignment, required=False)
-        self.metadata.add_input(Uri.TOKEN, required=False)
-        self.metadata.add_input(Uri.NE, required=False)
-        self.metadata.add_parameter('granularity', GRANULARITY_HELP, 'integer')
-        self.metadata.add_parameter('transcript', TRANSCRIPT_HELP, 'boolean')
-        return self.metadata
-
-    def _consume(self, mmif, **kwargs):
-        self.mmif = mmif if type(mmif) is Mmif else Mmif(mmif)
-        summary = Summary(mmif, **kwargs)
-        return summary.as_xml()
-        #print('!!! returning None for now')
+VERSION = '0.2.0'
+LICENSE = 'Apache 2.0'
 
 
 class Summary(object):
@@ -253,7 +172,7 @@ class Transcript(object):
                 # find the first text document in the Kaldi view
                 if anno.at_type.shortname == names.TEXT_DOCUMENT:
                     text = anno.properties['text'].value
-                    #text = f"{text} {text} {text}"
+                    # text = f"{text} {text} {text}"
                     lines = text.split('. ')
                     for line in lines[:-1]:
                         self.data.append(line + '.')
@@ -309,7 +228,7 @@ class SummarizedNodes(object):
 
 class SummarizedTimeFrames(SummarizedNodes):
 
-    """For now we take only the TimeFrames that have a frame type, which rules out
+    """For now, we take only the TimeFrames that have a frame type, which rules out
     all the frames we got from Kaldi."""
 
     def __init__(self, summary):
@@ -336,7 +255,7 @@ class SummarizedTimeFrames(SummarizedNodes):
 
 class SummarizedTags(SummarizedNodes):
 
-    """For now we take all semantic tags."""
+    """For now, we take all semantic tags."""
 
     def __init__(self, summary):
         super().__init__(summary)
@@ -368,45 +287,38 @@ class SummarizedEntities(SummarizedNodes):
         """Put all the entities from self.nodes in self.node_idx. This first puts
         the nodes into the dictionary indexed on text string and then sorts the
         list of nodes for each string on video position."""
-        for e in self:
-            self.nodes_idx.setdefault(e.properties['text'], []).append(e)
+        for ent in self:
+            self.nodes_idx.setdefault(ent.properties['text'], []).append(ent)
         for text, entities in self.nodes_idx.items():
             self.nodes_idx[text] = sorted(entities,
                                           key=(lambda e: e.start_in_video()))
 
     def _group(self):
         """Groups all the nodes on the text and sorts them on position in the video,
-        for the latter it will also created bins of entities that occur close to each
+        for the latter it will also create bins of entities that occur close to each
         other in the text."""
-        #print('>>> grouping entities, granularity=%s' % self.summary.granularity)
         # create the bins, governed by the summary's granularity
         self.bins = Bins(self.summary)
         for text, entities in self.nodes_idx.items():
-            #print('>>> adding %d instances of "%s"' % (len(entities), text))
             self.bins.current_bin = None
             for entity in entities:
                 self.bins.add_entity(text, entity)
         self.bins.mark_entities()
-        #self.bins.print_bins()
 
     def _add_tags(self, tags):
         for tag in tags:
             tag_doc = tag.properties['document']
             tag_p1 = tag.properties['start']
             tag_p2 = tag.properties['end']
-            #print('>>>', tag, tag_doc, tag_p1, tag_p2)
             entities = self.nodes_idx.get(tag.properties['text'], [])
             for entity in entities:
-                props = list(entity.properties.keys())
+                # props = list(entity.properties.keys())
                 props = entity.properties
                 doc = props['document']
                 p1 = props['start']
                 p2 = props['end']
-                #print('       ', entity, doc, p1, p2)
                 if tag_doc == doc and tag_p1 == p1 and tag_p2 == p2:
-                    #print('        ==> matches!')
                     entity.properties['tag'] = tag.properties['tagName']
-                    #print('           ', entity.properties)
 
     def as_xml(self):
         s = io.StringIO()
@@ -432,8 +344,6 @@ class SummarizedEntities(SummarizedNodes):
             print(key)
             for e in self.nodes_idx[key]:
                 print('   ', e, e.start_in_video())
-                #print('   ', e.paths_to_docs())
-                #print_paths(e.paths_to_docs(), indent='    ')
 
 
 class Bins(object):
@@ -442,6 +352,7 @@ class Bins(object):
         self.summary = summary
         self.bins = {}
         self.current_bin = None
+        self.current_text = None
 
     def add_entity(self, text, entity):
         """Add an entity instance to the appropriate bin."""
@@ -451,21 +362,16 @@ class Bins(object):
             self.current_text = text
             self.current_bin = Bin(entity)
             self.bins[text] = [self.current_bin]
-            #print('   ', entity, end='')
-            #print('  -- added new bin')
         else:
             # For following entities with the same text, a new bin may be
             # created depending on the positions and the granularity.
             p1 = self.current_bin[-1].start_in_video()
             p2 = entity.start_in_video()
-            p3 = entity.end_in_video()
-            #print('   ', entity, p1, p2, end='')
+            # p3 = entity.end_in_video()
             if p2 - p1 < self.summary.granularity:
-                #print('  -- added to current bin')
                 # TODO: should add p3 here
                 self.current_bin.add(entity)
             else:
-                #print('  -- added new bin')
                 self.current_bin = Bin(entity)
                 self.bins[self.current_text].append(self.current_bin)
 
@@ -475,8 +381,8 @@ class Bins(object):
         to be touched again."""
         # TODO: maybe use the bins when we create the output
         for entity_bins in self.bins.values():
-            for i, ebin in enumerate(entity_bins):
-                for entity in ebin:
+            for i, e_bin in enumerate(entity_bins):
+                for entity in e_bin:
                     entity.properties['group'] = i
 
     def print_bins(self):
@@ -510,24 +416,13 @@ class Bin(object):
             print(' ', i, node)
 
 
-def start_service():
-    summarizer = MmifSummarizer()
-    service = Restifier(summarizer, mimetype='application/xml')
-    service.run()
+if __name__ == '__main__':
 
-
-def test_on_sample(fname):
-    summarizer = MmifSummarizer()
-    result = summarizer.consume(open(fname).read(),
-                                granularity=GRANULARITY,
-                                transcript=True,
-                                transcript_mode='sentence')
-    print(result)
-
-
-if __name__ == "__main__":
-
-    if len(sys.argv) > 2 and sys.argv[1] == '-t':
-        test_on_sample(sys.argv[2])
-    else:
-        start_service()
+    fname = sys.argv[1]
+    with open(fname) as fh:
+        mmif_text = fh.read()
+        mmif_summary = Summary(mmif_text,
+                               granularity=GRANULARITY,
+                               transcript=True,
+                               transcript_mode='sentence')
+        print(mmif_summary.as_xml())
